+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * See here: https://github.com/breuner/elbencho/blob/master/docs/k8s-examples.md
 * Updated mimalloc lib for static builds to latest v3.3.2.
 * Wait 5sec after first CTRL+C press before resetting interrupt signal handler to have some time for graceful shutdown.
+* Keep result text/csv/json files in `/var/tmp` by default if no other path has been set.
 
 ### Contributors
 * Thanks to Daniel Drozdowski, Andrew Way, Xiaoyu Yao for helpful comments and suggestions.